adding traits for following a entity
Traits have been added to make the process of integrating following functionality to new entities more easily.

This means on the one hand you don't have to rewrite the same relationship code for followable entities over and over again, on the other hand there is only one place to add new followable entity relationshops on the followers side. Which will have a greater effect if there will be more then users that can follow an entity (for example if a company could follow an entity).